# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  staff: "staff",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  email?: string | null;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface Booking {
  id: number;
  guestName: string;
  phone: string;
  roomNumber: string;
  checkInTime: string;
}

export interface Room {
  id: number;
  number: string;
  status: RoomStatus;
  currentBooking?: Booking | null;
}

export interface CreateBookingRequest {
  guestName: string;
  phone: string;
  roomNumber: string;
}

export interface HistoryRecord {
  id: number;
  guestName: string;
  phone: string;
  roomNumber: string;
  checkInTime: string;
  checkOutTime: string;
}