# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.enum(["admin", "staff"]),
    email: zod.string().nullish(),
  }),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "staff"]),
  email: zod.string().nullish(),
});

/**
 * @summary Get all rooms with status
 */
export const GetRoomsResponseItem = zod.object({
  id: zod.number(),
  number: zod.string(),
  status: zod.enum(["available", "occupied"]),
  currentBooking: zod
    .object({
      id: zod.number(),
      guestName: zod.string(),
      phone: zod.string(),
      roomNumber: zod.string(),
      checkInTime: zod.string(),
    })
    .nullish(),
});
export const GetRoomsResponse = zod.array(GetRoomsResponseItem);

/**
 * @summary Get active bookings
 */
export const GetBookingsResponseItem = zod.object({
  id: zod.number(),
  guestName: zod.string(),
  phone: zod.string(),
  roomNumber: zod.string(),
  checkInTime: zod.string(),
});
export const GetBookingsResponse = zod.array(GetBookingsResponseItem);

/**
 * @summary Check in a guest
 */
export const CreateBookingBody = zod.object({
  guestName: zod.string(),
  phone: zod.string(),
  roomNumber: zod.string(),
});

/**
 * @summary Delete a booking (admin only)
 */
export const DeleteBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteBookingResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Check out a guest
 */
export const CheckoutBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const CheckoutBookingResponse = zod.object({
  id: zod.number(),
  guestName: zod.string(),
  phone: zod.string(),
  roomNumber: zod.string(),
  checkInTime: zod.string(),
  checkOutTime: zod.string(),
});

/**
 * @summary Get checkout history (last 2 months)
 */
export const GetHistoryResponseItem = zod.object({
  id: zod.number(),
  guestName: zod.string(),
  phone: zod.string(),
  roomNumber: zod.string(),
  checkInTime: zod.string(),
  checkOutTime: zod.string(),
});
export const GetHistoryResponse = zod.array(GetHistoryResponseItem);

/**
 * @summary Delete a history record (admin only)
 */
export const DeleteHistoryParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteHistoryResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});