# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  guestName: string;
  phone: string;
  roomNumber: string;
  checkInTime: string;
}