# lib/api-zod/src/generated/types/historyRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryRecord {
  id: number;
  guestName: string;
  phone: string;
  roomNumber: string;
  checkInTime: string;
  checkOutTime: string;
}