# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from "./booking";
import type { RoomStatus } from "./roomStatus";

export interface Room {
  id: number;
  number: string;
  status: RoomStatus;
  currentBooking?: Booking | null;
}